# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Madhuvastwa AI Marketing Automation API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /business:
    get:
      summary: List businesses for the authenticated owner
      tags: [Business]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of businesses
    post:
      summary: Create business profile
      tags: [Business]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Business'
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Business:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        targetAudience:
          type: string
        goals:
          type: string
        voiceTone:
          type: string
        brandColors:
          type: string
        logoUrl:
          type: string
        preferredEmotion:
          type: string
        preferredStyle:
          type: string
...